chore: rename files and fileds
- rename download/lib.js to download/libs/index.js
- rename `mirror` to `baseUri` in download/libs
